+++ Marduk.Controls/WaterfallFlowLayout.h
@@ -14,7 +14,7 @@ namespace Marduk
         {
         public:
             virtual RegisterReadOnlyProperty(double, _width, Width);
-            virtual RegisterReadOnlyProperty(Size, Size(Width, *std::max_element(_stacks->begin(), _stacks->end())), LayoutSize);
+            virtual RegisterReadOnlyProperty(Size, Size((float)Width, (float)*std::max_element(_stacks->begin(), _stacks->end())), LayoutSize);
             virtual RegisterReadOnlyProperty(int, _units->size(), UnitCount);
             RegisterReadOnlyProperty(double, _spacing, Spacing);
             RegisterReadOnlyProperty(int, _stacks->size(), StackCount);
